requirements.txt
Cython==0.29.14
PySDL2==0.9.6
